Replaces Table view for collection view
